backend/pyproject.toml
[project]
name = "fotomalovanky-admin"
version = "0.1.0"
description = "Order processing admin for Fotomalovanky.cz"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlmodel>=0.0.22",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "greenlet>=3.1.0",
    "dramatiq[redis]>=1.17.0",
    "httpx>=0.28.0",
    "pydantic-settings>=2.6.0",
    "structlog>=24.4.0",
    "pyjwt>=2.10.0",
    "psycopg[binary]>=3.2.0",
    "python-dateutil>=2.9.0",
    "python-dotenv>=1.0.0",
    "tenacity>=9.0.0",
    "pillow>=12.1.0",
    "python-ulid>=3.0.0",
    "aioboto3>=14.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-cov>=6.0.0",
    "pytest-asyncio>=0.24.0",
    "mypy>=1.13.0",
    "ruff>=0.8.0",
    "ariadne-codegen>=0.17.0",
    "types-python-dateutil>=2.9.0",
    "types-Pygments>=2.19.0",
    "types-aiobotocore[s3]>=2.15.0",
    "json5>=0.13.0",
    "pygments>=2.19.2",
    "click>=8.3.1",
]

[tool.uv]
default-groups = ["dev"]

[tool.ruff]
line-length = 120
target-version = "py314"
exclude = ["app/services/external/shopify_client/"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "T20", "RUF032"] # disabled "B" and "C90"; RUF032 detects Decimal() called with float literals
ignore = ["E203", "E231", "E266", "E501", "E701", "B904"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.mypy]
python_version = "3.14"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
exclude = ["app/services/external/shopify_client/"]

[[tool.mypy.overrides]]
module = "dramatiq.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "aioboto3.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "app.services.external.shopify_client.*"
ignore_errors = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v"

[project.scripts]
codegen = "app.scripts.codegen:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ariadne-codegen]
remote_schema_url = "https://aqi8it-7n.myshopify.com/admin/api/2025-01/graphql.json"
remote_schema_headers = { X-Shopify-Access-Token = "$SHOPIFY_ACCESS_TOKEN" }
queries_path = "app/graphql/queries"
target_package_path = "app/services/external/shopify_client"
async_client = true
client_name = "ShopifyClient"
plugins = ["ariadne_codegen.contrib.shorter_results.ShorterResultsPlugin"]
